pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "santa2025"
version = "0.1.0"
description = "Kaggle Santa 2025 team repository"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click",
  "hydra-core",
  "kaggle",
  "matplotlib",
  "numba",
  "numpy<2.0",
  "pandas",
  "psutil",
  "python-dotenv",
  "pyyaml",
  "scikit-learn",
  "seaborn",
  "shapely",
  "tqdm",
]

[dependency-groups]
dev = [
  "import-linter>=2.9",
  "marimo>=0.19.2",
  "pandas-stubs>=2.3.3.251219",
  "pre-commit",
  "pytest",
  "ruff",
  "ty>=0.0.11",
  "types-pyyaml>=6.0.12.20250915",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".import_linter_cache",
    ".venv",
    "third_party",
]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL", # Pylint
    "UP", # pyupgrade
]
ignore = [
    "B905",    # Zip-without-explicit-strict
    "F403",    # Unable to detect undefined names
    "N812",    # Lowercase imported as non-lowercase
    "N999",    # Invalid module name
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR2004", # Magic value used in comparison
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.importlinter]
root_package = "santa2025"

